fix compile error by moving PaymentTerm class (same as in master)
