+++ backend/de.metas.business/src/main/java/de/metas/payment/paymentterm/impl/PaymentTermRepository.java
@@ -5,10 +5,7 @@ import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.Maps;
 import de.metas.cache.CCache;
 import de.metas.organization.OrgId;
-import de.metas.payment.paymentterm.BaseLineType;
-import de.metas.payment.paymentterm.CalculationMethod;
-import de.metas.payment.paymentterm.IPaymentTermRepository;
-import de.metas.payment.paymentterm.PaymentTermId;
+import de.metas.payment.paymentterm.*;
 import de.metas.util.Services;
 import de.metas.util.lang.Percent;
 import lombok.NonNull;
