+++ backend/de.metas.business/src/main/java/de/metas/payment/paymentterm/interceptor/C_OrderLine.java
@@ -1,14 +1,11 @@
 package de.metas.payment.paymentterm.interceptor;
 
-import java.math.BigDecimal;
-
-import de.metas.payment.paymentterm.impl.PaymentTerm;
+import de.metas.payment.paymentterm.PaymentTerm;
 import org.adempiere.ad.callout.annotations.Callout;
 import org.adempiere.ad.callout.annotations.CalloutMethod;
 import org.adempiere.ad.callout.api.ICalloutField;
 import org.adempiere.ad.callout.spi.IProgramaticCalloutProvider;
 import org.compiere.model.I_C_OrderLine;
-import org.compiere.model.I_C_PaymentTerm;
 import org.springframework.stereotype.Component;
 
 import de.metas.order.IOrderLineBL;
