+++ backend/de.metas.externalsystem/src/main/java/de/metas/externalsystem/process/InvokeShopwareService.java
@@ -43,14 +43,13 @@ import de.metas.externalsystem.shopware6.UOMShopwareMapping;
 import de.metas.order.impl.DocTypeService;
 import de.metas.payment.paymentterm.IPaymentTermRepository;
 import de.metas.payment.paymentterm.PaymentTermId;
-import de.metas.payment.paymentterm.impl.PaymentTerm;
+import de.metas.payment.paymentterm.PaymentTerm;
 import de.metas.pricing.PriceListId;
 import de.metas.pricing.service.IPriceListDAO;
 import de.metas.uom.IUOMDAO;
 import de.metas.util.Services;
 import lombok.NonNull;
 import org.adempiere.exceptions.AdempiereException;
-import org.compiere.model.I_C_PaymentTerm;
 import org.compiere.model.I_M_PriceList;
 import org.springframework.stereotype.Service;
